refactor: split now-playing.command.ts into focused modules
Extract leaf helpers and the journal flow out of the 3068-line
NowPlayingCommand into dedicated modules under src/commands/now-playing/:
modals, component builders, channel utils, IGDB import helpers, and a
NowPlayingJournalHandlers @Discord class. Command file drops to ~2193 lines.
Custom-id strings/regexes and as-any casts moved verbatim; the full
component/slash decorator set is unchanged (46 handlers).

Closes #892
